requirements.txt
alembic==1.7.7
colorama==0.4.5
greenlet==2.0.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.50
typing-extensions==4.1.1
watchdog==2.3.1
zipp==3.6.0
